test(web): complete session検証を純粋関数へ移す
Summary:
- completeInterviewSession の保存 payload 生成を shared utils の純粋関数へ
  切り出した。
- repository と moderation service を vi.mock していた service test を削除し、
  自動公開フラグと根拠メッセージ補完を純粋関数テストで検証した。

Rationale:
- PR レビューで指摘された complete-interview-session test の mock をなくし、
  server-only や外部 API を含まない形で保存判定ロジックを検証するため。
- 実 DB 経由の integration test は既存の service 完了フローを引き続き
  カバーし、判定ロジックは単体で決定論的に確認できるようにするため。

Tests:
- pnpm --filter web test -- complete-interview-report
- pnpm --filter web typecheck
- pnpm --filter web test:integration -- complete-interview-session public-report-loaders
- pnpm lint
- pnpm typecheck
- pnpm build
- pnpm test
- pnpm --filter web test:coverage
- git diff --check
